# Speedup SpanHelpers.IndexOf{Any}(byte, ...)