--- a/src/System.Private.CoreLib/shared/System/SpanHelpers.cs
+++ b/src/System.Private.CoreLib/shared/System/SpanHelpers.cs
@@ -3,10 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Diagnostics;
-using System.Globalization;
 using System.Runtime;
-using System.Runtime.CompilerServices;
-using System.Runtime.Intrinsics.X86;
 
 using Internal.Runtime.CompilerServices;