Learnings from https://github.com/dotnet/coreclr/pull/22019; improvement on https://github.com/dotnet/coreclr/pull/21073

Applied to
```csharp
internal static partial class SpanHelpers
{
    static int IndexOf(ref byte searchSpace, byte value, int length);
    static int IndexOfAny(ref byte searchSpace, byte value0, byte value1, int length);
    static int IndexOfAny(ref byte searchSpace, byte value0, byte value1, byte value2, int length);
}
```


`MoveMask`/`vpmovmskb` used for equality in the vectorized path already has flagged the bits that match; so if we use the specific intrinsics rather than generic Vector, we just need to determine the bit set, rather than do further processing to determine the element offset.

```diff
-; Total bytes of code 597, prolog size 5 for method SpanHelpers:IndexOf(byref,ubyte,int):int
+; Total bytes of code 549, prolog size 5 for method SpanHelpers:IndexOf(byref,ubyte,int):int
```

Performance measurements:

Array length 512 significant improvements (up to more than double) for found item in any position. https://github.com/dotnet/coreclr/pull/22118#issuecomment-456199510 

Array lengths >= 32 with item in last position significant improvements  (up to more than double)  https://github.com/dotnet/coreclr/pull/22118#issuecomment-456223461

/cc @CarolEidt @fiigii @tannergooding @ahsonkhan