@benaadams 2019-01-21
```diff
-; Total bytes of code 597, prolog size 5 for method SpanHelpers:IndexOf(byref,ubyte,int):int
+; Total bytes of code 546, prolog size 5 for method SpanHelpers:IndexOf(byref,ubyte,int):int
```