@benaadams 2019-01-21
```
 Method   | Pos |    Current |         PR | % Improvement |
----------|-----|------------|------------|---------------|
 IndexOf  |   0	|   6.403 ns |   5.307 ns |       +20.65% |
 IndexOf  |   1	|   6.256 ns |   5.506 ns |       +13.62% |
 IndexOf  |   2	|   6.192 ns |   5.501 ns |       +12.56% |
 IndexOf  |   3	|   6.425 ns |   5.498 ns |       +16.86% |
 IndexOf  |   4	|   6.394 ns |   5.401 ns |       +18.38% |
 IndexOf  |   7	|   6.176 ns |   5.322 ns |       +16.04% |
 IndexOf  |   8	|   6.265 ns |   5.390 ns |       +16.23% |
 IndexOf  |   9	|   6.202 ns |   5.538 ns |       +11.98% |
 IndexOf  |  10	|   6.274 ns |   5.430 ns |       +15.54% |
 IndexOf  |  11	|   6.198 ns |   5.518 ns |       +12.32% |
 IndexOf  |  12	|   6.537 ns |   5.873 ns |       +11.30% |
 IndexOf  |  13	|   7.124 ns |   6.593 ns |        +8.05% |
 IndexOf  |  14	|   7.530 ns |   6.522 ns |       +15.45% |
 IndexOf  |  15	|   7.055 ns |   6.576 ns |        +7.28% |
 IndexOf  |  16	|   9.521 ns |   8.457 ns |       +12.58% |
 IndexOf  |  17	|   9.510 ns |   8.586 ns |       +10.76% |
 IndexOf  |  18	|   9.452 ns |   8.613 ns |        +9.74% |
 IndexOf  |  19	|   9.518 ns |   8.492 ns |       +12.08% |
 IndexOf  |  20	|   9.511 ns |   8.472 ns |       +12.26% |
 IndexOf  |  21	|   9.501 ns |   8.493 ns |       +11.86% |
 IndexOf  |  22	|   9.458 ns |   8.466 ns |       +11.71% |
 IndexOf  |  23	|   9.278 ns |   8.691 ns |        +6.75% |
 IndexOf  |  24	|   9.904 ns |   8.593 ns |       +15.25% |
 IndexOf  |  25	|   9.697 ns |   8.692 ns |       +11.56% |
 IndexOf  |  26	|   9.720 ns |   8.656 ns |       +12.29% |
 IndexOf  |  27	|   9.819 ns |   8.485 ns |       +15.72% |
 IndexOf  |  28	|   9.795 ns |   8.642 ns |       +13.34% |
 IndexOf  |  29	|   9.584 ns |   8.548 ns |       +12.11% |
 IndexOf  |  30	|   9.751 ns |   8.501 ns |       +14.70% |
 IndexOf  |  31	|   9.778 ns |   8.533 ns |       +14.59% |
 IndexOf  |  32	|  10.110 ns |   8.398 ns |       +20.38% |
 IndexOf  |  33	|   9.933 ns |   8.593 ns |       +15.59% |
 IndexOf  |  34	|   9.922 ns |   8.599 ns |       +15.38% |
 IndexOf  |  35	|  10.034 ns |   8.476 ns |       +18.38% |
 IndexOf  |  36	|   9.994 ns |   8.434 ns |       +18.49% |
 IndexOf  |  37	|  10.059 ns |   8.564 ns |       +17.45% |
 IndexOf  |  38	|  10.101 ns |   8.630 ns |       +17.04% |
 IndexOf  |  39	|  10.096 ns |   8.332 ns |       +21.17% |
 IndexOf  |  40	|  10.147 ns |   8.688 ns |       +16.79% |
 IndexOf  |  41	|  10.449 ns |   8.332 ns |       +25.40% |
 IndexOf  |  42	|  10.161 ns |   8.690 ns |       +16.92% |
 IndexOf  |  43	|  10.296 ns |   8.546 ns |       +20.47% |
 IndexOf  |  44	|  10.240 ns |   8.662 ns |       +18.21% |
 IndexOf  |  45	|  10.385 ns |   8.575 ns |       +21.10% |
 IndexOf  |  46	|  10.246 ns |   8.406 ns |       +21.88% |
 IndexOf  |  47	|  10.427 ns |   8.661 ns |       +20.39% |
 IndexOf  |  48	|  10.877 ns |   9.998 ns |        +8.79% |
 IndexOf  |  49	|  10.688 ns |  10.046 ns |        +6.39% |
 IndexOf  |  50	|  10.895 ns |  10.299 ns |        +5.78% |
 IndexOf  |  51	|  10.919 ns |  10.072 ns |        +8.40% |
 IndexOf  |  52	|  10.591 ns |  10.227 ns |        +3.55% |
 IndexOf  |  53	|  10.772 ns |  10.277 ns |        +4.81% |
 IndexOf  |  54	|  10.760 ns |  10.258 ns |        +4.89% |
 IndexOf  |  55	|  10.898 ns |  10.202 ns |        +6.82% |
 IndexOf  |  56	|  10.776 ns |  10.044 ns |        +7.28% |
 IndexOf  |  57	|  11.472 ns |  10.269 ns |       +11.71% |
 IndexOf  |  58	|  10.756 ns |  10.217 ns |        +5.27% |
 IndexOf  |  59	|  10.946 ns |  10.254 ns |        +6.74% |
 IndexOf  |  60	|  10.745 ns |  10.278 ns |        +4.54% |
 IndexOf  |  61	|  10.889 ns |  10.241 ns |        +6.32% |
 IndexOf  |  62	|  10.755 ns |  10.241 ns |        +5.01% |
 IndexOf  |  63	|  10.777 ns |  10.286 ns |        +4.77% |
 IndexOf  |  64	|  11.007 ns |  10.196 ns |        +7.95% |
 IndexOf  |  65	|  11.187 ns |  10.304 ns |        +8.56% |
 IndexOf  |  66	|  11.009 ns |  10.276 ns |        +7.13% |
 IndexOf  |  67	|  11.129 ns |  10.280 ns |        +8.25% |
 IndexOf  |  68	|  11.129 ns |   9.933 ns |       +12.04% |
 IndexOf  |  69	|  10.839 ns |  10.280 ns |        +5.43% |
 IndexOf  |  70	|  11.207 ns |  10.285 ns |        +8.96% |
 IndexOf  |  71	|  11.139 ns |  10.032 ns |       +11.03% |
 IndexOf  |  72	|  11.345 ns |  10.009 ns |       +13.34% |
 IndexOf  |  73	|  11.559 ns |  10.250 ns |       +12.77% |
 IndexOf  |  74	|  11.396 ns |  10.121 ns |       +12.59% |
 IndexOf  |  75	|  11.271 ns |  10.259 ns |        +9.86% |
 IndexOf  |  76	|  11.282 ns |  10.177 ns |       +10.85% |
 IndexOf  |  77	|  11.889 ns |  10.114 ns |       +17.54% |
 IndexOf  |  78	|  11.292 ns |  10.159 ns |       +11.15% |
 IndexOf  |  79	|  11.417 ns |  10.046 ns |       +13.64% |
 IndexOf  |  80	|  11.843 ns |  10.800 ns |        +9.65% |
 IndexOf  |  81	|  11.679 ns |  10.416 ns |       +12.12% |
 IndexOf  |  82	|  11.755 ns |  10.548 ns |       +11.44% |
 IndexOf  |  83	|  11.757 ns |  10.626 ns |       +10.64% |
 IndexOf  |  84	|  11.785 ns |  10.856 ns |        +8.55% |
 IndexOf  |  85	|  11.625 ns |  10.707 ns |        +8.57% |
 IndexOf  |  86	|  11.900 ns |  11.184 ns |        +6.40% |
 IndexOf  | 126	|  10.102 ns |   7.931 ns |       +27.37% |
 IndexOf  | 127	|  10.030 ns |   6.433 ns |       +55.91% |
 IndexOf  | 128	|  10.144 ns |   7.150 ns |       +41.87% |
 IndexOf  | 129	|  10.342 ns |   7.126 ns |       +45.13% |
 IndexOf  | 130	|  10.213 ns |   7.358 ns |       +38.80% |
 IndexOf  | 131	|  10.170 ns |   7.324 ns |       +38.85% |
 IndexOf  | 250	|  13.517 ns |   8.269 ns |       +63.46% |
 IndexOf  | 251	|  13.416 ns |   8.283 ns |       +61.97% |
 IndexOf  | 252	|  13.473 ns |   8.383 ns |       +60.71% |
 IndexOf  | 253	|  13.629 ns |  10.251 ns |       +32.95% |
 IndexOf  | 254	|  13.355 ns |   8.101 ns |       +64.85% |
 IndexOf  | 255	|  26.023 ns |  21.101 ns |       +23.32% |
```