@benaadams 2019-01-21
Some interesting speed bumps in there; might be able to go better