@benaadams 2019-01-21
```csharp
public class Program
{
    private static byte[] s_source;
    private const int Iters = 100;

    public static void Main(string[] args)
    {
        var summary = BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
    }

    [Params(
        0, 1, 2, 3, 4, 7, 
        8, 9, 10, 11, 12, 13, 14, 15, 
        16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 
        32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 
        48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 
        64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
        80, 81, 82, 83, 84, 85, 86,
        126, 127, 128, 129, 130, 131,
        250, 251, 252, 253, 254, 255)]
    public int Position { get; set; }

    [Benchmark(OperationsPerInvoke = Iters)]
    public int IndexOf()
    {
        int total = 0;
        byte value = (byte)Position;
        ReadOnlySpan<byte> span = s_source;
        for (int i = 0; i < Iters; i++)
        {
            total += span.IndexOf(value);
        }

        return total;
    }

    [GlobalSetup]
    public void Setup()
    {
        var e = Enumerable.Range(0, 255).Select(b => (byte)b);
        s_source = e.Concat(e).ToArray();
    }
}
```

```
 Method   | Pos |    Current |         PR | % Improvement |
----------|-----|------------|------------|---------------|
  IndexOf |   0 |   6.403 ns |   4.461 ns |        +43.5% |
  IndexOf |   1 |   6.256 ns |   4.516 ns |        +38.5% |
  IndexOf |   2 |   6.192 ns |   4.494 ns |        +37.7% |
  IndexOf |   3 |   6.425 ns |   4.460 ns |        +44.0% |
  IndexOf |   4 |   6.394 ns |   4.420 ns |        +44.6% |
  IndexOf |   7 |   6.176 ns |   4.424 ns |        +39.6% |
  IndexOf |   8 |   6.265 ns |   4.464 ns |        +40.3% |
  IndexOf |   9 |   6.202 ns |   4.564 ns |        +35.8% |
  IndexOf |  10 |   6.274 ns |   4.458 ns |        +40.7% |
  IndexOf |  11 |   6.198 ns |   4.607 ns |        +34.5% |
  IndexOf |  12 |   6.537 ns |   4.469 ns |        +46.2% |
  IndexOf |  13 |   7.124 ns |   4.421 ns |        +61.1% |
  IndexOf |  14 |   7.530 ns |   4.429 ns |        +70.0% |
  IndexOf |  15 |   7.055 ns |   4.401 ns |        +60.3% |
  IndexOf |  16 |   9.521 ns |   4.421 ns |       +115.3% |
  IndexOf |  17 |   9.510 ns |   4.863 ns |        +95.5% |
  IndexOf |  18 |   9.452 ns |   4.531 ns |       +108.6% |
  IndexOf |  19 |   9.518 ns |   4.535 ns |       +109.8% |
  IndexOf |  20 |   9.511 ns |   4.499 ns |       +111.4% |
  IndexOf |  21 |   9.501 ns |   4.322 ns |       +119.8% |
  IndexOf |  22 |   9.458 ns |   4.418 ns |       +114.0% |
  IndexOf |  23 |   9.278 ns |   4.579 ns |       +102.6% |
  IndexOf |  24 |   9.904 ns |   4.637 ns |       +113.5% |
  IndexOf |  25 |   9.697 ns |   4.667 ns |       +107.7% |
  IndexOf |  26 |   9.720 ns |   4.459 ns |       +117.9% |
  IndexOf |  27 |   9.819 ns |   4.380 ns |       +124.1% |
  IndexOf |  28 |   9.795 ns |   4.520 ns |       +116.7% |
  IndexOf |  29 |   9.584 ns |   4.628 ns |       +107.0% |
  IndexOf |  30 |   9.751 ns |   4.457 ns |       +118.7% |
  IndexOf |  31 |   9.778 ns |   4.520 ns |       +116.3% |
  IndexOf |  32 |  10.110 ns |   6.077 ns |        +66.3% |
  IndexOf |  33 |   9.933 ns |   6.129 ns |        +62.0% |
  IndexOf |  34 |   9.922 ns |   6.183 ns |        +60.4% |
  IndexOf |  35 |  10.034 ns |   6.155 ns |        +63.0% |
  IndexOf |  36 |   9.994 ns |   5.244 ns |        +90.5% |
  IndexOf |  37 |  10.059 ns |   6.211 ns |        +61.9% |
  IndexOf |  38 |  10.101 ns |   5.085 ns |        +98.6% |
  IndexOf |  39 |  10.096 ns |   6.255 ns |        +61.4% |
  IndexOf |  40 |  10.147 ns |   5.560 ns |        +82.5% |
  IndexOf |  41 |  10.449 ns |   6.150 ns |        +69.9% |
  IndexOf |  42 |  10.161 ns |   5.542 ns |        +83.3% |
  IndexOf |  43 |  10.296 ns |   5.075 ns |       +102.8% |
  IndexOf |  44 |  10.240 ns |   4.993 ns |       +105.0% |
  IndexOf |  45 |  10.385 ns |   4.956 ns |       +109.5% |
  IndexOf |  46 |  10.246 ns |   6.147 ns |        +66.6% |
  IndexOf |  47 |  10.427 ns |   4.988 ns |       +109.0% |
  IndexOf |  48 |  10.877 ns |   6.229 ns |        +74.6% |
  IndexOf |  49 |  10.688 ns |   6.090 ns |        +75.5% |
  IndexOf |  50 |  10.895 ns |   6.019 ns |        +81.0% |
  IndexOf |  51 |  10.919 ns |   6.206 ns |        +75.9% |
  IndexOf |  52 |  10.591 ns |   6.012 ns |        +76.1% |
  IndexOf |  53 |  10.772 ns |   6.232 ns |        +72.8% |
  IndexOf |  54 |  10.760 ns |   6.212 ns |        +73.2% |
  IndexOf |  55 |  10.898 ns |   5.241 ns |       +107.9% |
  IndexOf |  56 |  10.776 ns |   4.983 ns |       +116.2% |
  IndexOf |  57 |  11.472 ns |   6.150 ns |        +86.5% |
  IndexOf |  58 |  10.756 ns |   7.093 ns |        +51.6% |
  IndexOf |  59 |  10.946 ns |   4.891 ns |       +123.7% |
  IndexOf |  60 |  10.745 ns |   5.260 ns |       +104.2% |
  IndexOf |  61 |  10.889 ns |   6.158 ns |        +76.8% |
  IndexOf |  62 |  10.755 ns |   6.050 ns |        +77.7% |
  IndexOf |  63 |  10.777 ns |   4.971 ns |       +116.7% |
  IndexOf |  64 |  11.007 ns |   5.782 ns |        +90.3% |
  IndexOf |  65 |  11.187 ns |   6.599 ns |        +69.5% |
  IndexOf |  66 |  11.009 ns |   5.690 ns |        +93.4% |
  IndexOf |  67 |  11.129 ns |   5.653 ns |        +96.8% |
  IndexOf |  68 |  11.129 ns |   5.775 ns |        +92.7% |
  IndexOf |  69 |  10.839 ns |   6.633 ns |        +63.4% |
  IndexOf |  70 |  11.207 ns |   5.734 ns |        +95.4% |
  IndexOf |  71 |  11.139 ns |   5.611 ns |        +98.5% |
  IndexOf |  72 |  11.345 ns |   5.701 ns |        +99.0% |
  IndexOf |  73 |  11.559 ns |   6.728 ns |        +71.8% |
  IndexOf |  74 |  11.396 ns |   5.614 ns |       +102.9% |
  IndexOf |  75 |  11.271 ns |   6.672 ns |        +68.9% |
  IndexOf |  76 |  11.282 ns |   6.741 ns |        +67.3% |
  IndexOf |  77 |  11.889 ns |   5.737 ns |       +107.2% |
  IndexOf |  78 |  11.292 ns |   6.755 ns |        +67.1% |
  IndexOf |  79 |  11.417 ns |   6.746 ns |        +69.2% |
  IndexOf |  80 |  11.843 ns |   5.734 ns |       +106.5% |
  IndexOf |  81 |  11.679 ns |   6.817 ns |        +71.3% |
  IndexOf |  82 |  11.755 ns |   5.670 ns |       +107.3% |
  IndexOf |  83 |  11.757 ns |   5.736 ns |       +104.9% |
  IndexOf |  84 |  11.785 ns |   5.743 ns |       +105.2% |
  IndexOf |  85 |  11.625 ns |   6.828 ns |        +70.2% |
  IndexOf |  86 |  11.900 ns |   6.753 ns |        +76.2% |
  IndexOf | 126 |  10.102 ns |   6.467 ns |        +56.2% |
  IndexOf | 127 |  10.030 ns |   6.393 ns |        +56.8% |
  IndexOf | 128 |  10.144 ns |   7.130 ns |        +42.2% |
  IndexOf | 129 |  10.342 ns |   7.978 ns |        +29.6% |
  IndexOf | 130 |  10.213 ns |   8.016 ns |        +27.4% |
  IndexOf | 131 |  10.170 ns |   7.112 ns |        +42.9% |
  IndexOf | 250 |  13.517 ns |   9.686 ns |        +39.5% |
  IndexOf | 251 |  13.416 ns |   9.778 ns |        +37.2% |
  IndexOf | 252 |  13.473 ns |   7.693 ns |        +75.1% |
  IndexOf | 253 |  13.629 ns |   9.824 ns |        +38.7% |
  IndexOf | 254 |  13.355 ns |   9.702 ns |        +37.6% |
  IndexOf | 255 |  26.023 ns |  22.202 ns |        +17.2% |
```